+++ Changelog.md
@@ -40,6 +40,7 @@ Bugfixes:
 
  * JSON AST: nodes were added at wrong parent
  * Why3 translator: crash fix for exponentiation
+ * Commandline Interface: linking libraries with underscores in their name.
  * Type Checker: Fallback function cannot return data anymore.
  * Code Generator: Fix crash when sha3() was used on unsupported types.